Move OnWrapAction to shared system
